+++ projects/iplab/ngx-file-upload/src/lib/directives/attribute.directive.ts
@@ -1,4 +1,4 @@
-import { Directive, Input, SimpleChanges, Optional, Host, Self, OnChanges, AfterViewInit } from '@angular/core';
+import { Directive, Input, SimpleChanges, Optional, Host, Self, OnChanges, AfterViewInit, input, InputSignal, effect } from '@angular/core';
 import { FileUploadComponent } from '../components/multiple-file-upload/file-upload.component';
 import { SimpleFileUploadComponent } from '../components/simple-file-upload/simple-file-upload.component';
 
@@ -6,13 +6,12 @@ import { SimpleFileUploadComponent } from '../components/simple-file-upload/simp
 
 @Directive({
     selector: 'file-upload[accept]',
-    host: {'[attr.accept]': 'accept ? accept : null'},
+    host: { '[attr.accept]': 'accept ? accept : null' },
     standalone: true
 })
-export class FilesAcceptDirective implements AfterViewInit, OnChanges {
+export class FilesAcceptDirective {
 
-    @Input()
-    public accept: string;
+    public accept: InputSignal<string> = input.required();
 
     private readonly fileUpload: FileUploadComponent | SimpleFileUploadComponent = null;
 
@@ -20,21 +19,15 @@ export class FilesAcceptDirective implements AfterViewInit, OnChanges {
         @Optional() @Host() @Self() fileUpload: FileUploadComponent,
         @Optional() @Host() @Self() simpleFileUpload: SimpleFileUploadComponent) {
         this.fileUpload = fileUpload || simpleFileUpload;
-    }
-
-    public ngAfterViewInit(): void {
-        this.setAccept(this.accept);
-    }
-
-    public ngOnChanges(changes: SimpleChanges): void {
-        if ('accept' in changes && changes['accept'].currentValue !== changes['accept'].previousValue) {
-            this.setAccept(this.accept);
-        }
+        effect(() => {
+            this.setAccept(this.accept());
+        });
     }
 
     private setAccept(accept: string): void {
-        if (this.fileUpload && this.fileUpload.control) {
-            this.fileUpload.control.acceptFiles(accept);
+        const control = this.fileUpload?.control();
+        if (control) {
+            control.acceptFiles(accept);
         }
     }
 }