+++ projects/iplab/ngx-file-upload/src/lib/directives/discard.directive.ts
@@ -1,28 +1,27 @@
-import { Directive, Input, SimpleChanges, Optional, Host, Self, OnChanges, AfterViewInit } from '@angular/core';
+import { Directive, Input, SimpleChanges, Optional, Host, Self, OnChanges, AfterViewInit, input, InputSignalWithTransform, booleanAttribute, effect } from '@angular/core';
 import { FileUploadComponent } from '../components/multiple-file-upload/file-upload.component';
 import { SimpleFileUploadComponent } from '../components/simple-file-upload/simple-file-upload.component';
 
 
+export function booleanTransform(value: unknown): boolean {
+    if (typeof value === 'string' && (value === 'true' || value === 'false')) {
+        return JSON.parse(value.toLowerCase());
+    } else if (typeof value === 'boolean') {
+        return value;
+    } else {
+        throw Error(`Provided value in directive [discard]="${value}" is not boolean.`);
+    }
+}
+
 
 @Directive({
     selector: 'file-upload[discard]',
-    host: {'[attr.discard]': 'discard ? discard : null'},
+    host: { '[attr.discard]': 'discard ? discard : null' },
     standalone: true
 })
-export class FilesDiscardDirective implements AfterViewInit, OnChanges {
+export class FilesDiscardDirective {
 
-    private discardValue: boolean | null = null;
-
-    @Input()
-    public set discard(discard: boolean | string) {
-        if (typeof discard === 'string' && (discard === 'true' || discard === 'false')) {
-            this.discardValue = JSON.parse(discard.toLowerCase());
-        } else if (typeof discard === 'boolean') {
-            this.discardValue = discard;
-        } else {
-            throw Error(`Provided value in directive [discard]="${discard}" is not boolean.`);
-        }
-    }
+    public discard: InputSignalWithTransform<boolean, boolean | string | null> = input<boolean, boolean | string | null>(true, { transform: booleanAttribute });
 
     private readonly fileUpload: FileUploadComponent | SimpleFileUploadComponent = null;
 
@@ -30,21 +29,15 @@ export class FilesDiscardDirective implements AfterViewInit, OnChanges {
         @Optional() @Host() @Self() fileUpload: FileUploadComponent,
         @Optional() @Host() @Self() simpleFileUpload: SimpleFileUploadComponent) {
         this.fileUpload = fileUpload || simpleFileUpload;
-    }
-
-    public ngAfterViewInit(): void {
-        this.setAccept(this.discardValue);
-    }
-
-    public ngOnChanges(changes: SimpleChanges): void {
-        if ('discard' in changes && changes['discard'].currentValue !== changes['discard'].previousValue) {
-            this.setAccept(this.discardValue);
-        }
+        effect(() => {
+            this.setAccept(this.discard());
+        });
     }
 
     private setAccept(discard: boolean): void {
-        if (this.fileUpload && this.fileUpload.control) {
-            this.fileUpload.control.discardInvalid(discard);
+        const control = this.fileUpload?.control();
+        if (control) {
+            control.discardInvalid(discard);
         }
     }
 }