+++ src/main.ts
@@ -1,15 +1,20 @@
-import { enableProdMode } from '@angular/core';
+import { enableProdMode, provideExperimentalZonelessChangeDetection } from '@angular/core';
 import { bootstrapApplication } from '@angular/platform-browser';
-import { provideAnimations, provideNoopAnimations } from '@angular/platform-browser/animations';
+// import { provideAnimations, provideNoopAnimations } from '@angular/platform-browser/animations';
+import { provideAnimationsAsync } from '@angular/platform-browser/animations/async';
 
 import { environment } from './environments/environment';
 import { AppComponent } from './example-app/app.component';
+
 
 if (environment.production) {
   enableProdMode();
 }
 
 bootstrapApplication(AppComponent, {
-  providers: provideAnimations()
+  providers: [
+    provideAnimationsAsync('animations')
+    ,provideExperimentalZonelessChangeDetection()
+  ]
 })
 .catch(err => console.log(err));