+++ CHANGELOG.md
@@ -1,4 +1,12 @@
 
+<a name="18.1.0"></a>
+## [18.0.1](https://github.com/pIvan/file-upload/commit/82cf9fc3d613ea563f72872b204416f58dee997b) (2024-06-20)
+
+### Bug Fix
+* **Value changes** - preventing valueChanges event emit on model->view update
+### Features
+* **signals** - New signals implemented
+
 <a name="18.0.0"></a>
 ## [18.0.0](https://github.com/pIvan/file-upload/commit/82cf9fc3d613ea563f72872b204416f58dee997b) (2024-06-12)